some structs moved to MergeReport project, because not used anywhere else
