+++ Core/runfrml.cpp
@@ -7,12 +7,13 @@
 #include "../Common/compare.h"
 #include "../Common/pstring.h"
 #include "../Common/textfunc.h"
-#include "..\DataEditor\DataEditor.h"
+#include "../DataEditor/DataEditor.h"
 #include "../fandio/directory.h"
 #include "../fandio/FandTFile.h"
 #include "../fandio/XWKey.h"
 #include "../pascal/random.h"
 #include "../Drivers/constants.h"
+#include "../MergeReport/InpD.h"
 
 #include "FieldDescr.h"
 #include "FileD.h"
