+++ MergeReport/ConstListEl.h
@@ -1,0 +1,9 @@
+#pragma once
+#include <string>
+
+struct ConstListEl
+{
+	std::string S;
+	double R = 0;
+	bool B = false;
+};
