+++ MergeReport/Report.cpp
@@ -1,5 +1,9 @@
 #include "Report.h"
+
+#include "ConstListEl.h"
+#include "InpD.h"
 #include "MergeReportBase.h"
+#include "RFldD.h"
 #include "../Common/compare.h"
 #include "../Common/textfunc.h"
 #include "../Core/Compiler.h"
